Disable the engine watchdog timer while in debug shell
There's a watchdog timer on linux checking if the server is taking too long to process a frame. It can be disabled using `-nowatchdog` or tamed using `-nonabortingwatchdog` in the server command line. Take care of it just while the debug shell is active and keep it enabled while the server is running normally.