Remove weird breakpoint spam protection
There is special handling of hitting multiple breakpoints on a single line in pawndbg, which should prevent getting stuck on a line with multiple statements? I'd like to be able to step through the statements, so remove that protection.

It also prevented a newly created breakpoint from being hit if the plugin was continued right away without stepping to a the next line first.

```
break .
continue
```
The breakpoint wouldn't be hit X times before the protection would run off.